backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==1.10.13
python-decouple==3.8

# Database
pymongo==4.6.0
motor==3.3.1
beanie==1.23.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyjwt==2.8.0

# AI & ML Services
openai==1.3.5
boto3==1.29.6

# Audio Processing
whisper==1.1.10
ffmpeg-python==0.2.0

# File Processing
python-magic==0.4.27
python-docx==1.0.0
PyPDF2==3.0.1
docx2txt==0.8

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-slugify==8.0.1
email-validator==2.1.0.post1
pytz==2023.3.post1
tenacity==8.2.3

# Monitoring & Logging
prometheus-fastapi-instrumentator<6.0.0
structlog==23.2.0

# CORS
starlette==0.27.0
starlette-csrf==2.0.0

# Rate Limiting
fastapi-limiter==0.1.5
redis==4.5.5

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 